+++ example/lib/example5_themes_playground/widgets/panels/card_and_material_settings/card_settings.dart
@@ -74,7 +74,7 @@ class CardSettings extends StatelessWidget {
                           : 'default 4',
                   style: Theme.of(context)
                       .textTheme
-                      .caption!
+                      .bodySmall!
                       .copyWith(fontWeight: FontWeight.bold),
                 ),
               ],