+++ example/lib/example5_themes_playground/widgets/panels/card_and_material_settings/material_and_bottom_sheet_settings.dart
@@ -79,7 +79,7 @@ class MaterialAndBottomSheetSettings extends StatelessWidget {
                       : 'default 0',
                   style: Theme.of(context)
                       .textTheme
-                      .caption!
+                      .bodySmall!
                       .copyWith(fontWeight: FontWeight.bold),
                 ),
               ],