+++ example/lib/example5_themes_playground/widgets/panels/navigation_bar_settings/bottom_navigation_bar_settings.dart
@@ -103,7 +103,7 @@ class BottomNavigationBarSettings extends StatelessWidget {
                   '${(navBarOpacity * 100).toStringAsFixed(0)} %',
                   style: Theme.of(context)
                       .textTheme
-                      .caption!
+                      .bodySmall!
                       .copyWith(fontWeight: FontWeight.bold),
                 ),
               ],
@@ -135,7 +135,7 @@ class BottomNavigationBarSettings extends StatelessWidget {
                   navBarElevation.toStringAsFixed(1),
                   style: Theme.of(context)
                       .textTheme
-                      .caption!
+                      .bodySmall!
                       .copyWith(fontWeight: FontWeight.bold),
                 ),
               ],