+++ example/lib/example5_themes_playground/widgets/panels/toggle_buttons_settings/toggle_buttons_settings.dart
@@ -90,7 +90,7 @@ class ToggleButtonsSettings extends StatelessWidget {
                       : 'default 0',
                   style: Theme.of(context)
                       .textTheme
-                      .caption!
+                      .bodySmall!
                       .copyWith(fontWeight: FontWeight.bold),
                 ),
               ],
@@ -147,7 +147,7 @@ class ToggleButtonsSettings extends StatelessWidget {
                       : 'default 1',
                   style: Theme.of(context)
                       .textTheme
-                      .caption!
+                      .bodySmall!
                       .copyWith(fontWeight: FontWeight.bold),
                 ),
               ],